Intro to AI Practicum Project
This commit contains the unity project for the Intro to AI Practicum. It adds the ML Agents library and the RLBattleAgent script which contains the code for an ML Agent that can control the first player. It also makes changes to the BattleCore to allow the a battle agent to pass in inputs.