MLFootsies Project - Final Push
This push contains the final state of the MLFootsies project for machine learning. This push contains all source code, game assets, and ML-agents packages and training data needed to compile and run the game in version 2021.3.6f1 of the Unity game engine. This push also contains a final windows build of the game, Footsies.exe, that allows the user to play against an AI trained through 5,000,000 episodes of adversarial training using the ML-agents library.